% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_track_intersection.R
\name{cross_track_intersection}
\alias{cross_track_intersection}
\title{Calculate cross-track intersection}
\usage{
cross_track_intersection(b, a1, a2)
}
\arguments{
\item{b}{coordinates of the body, b:
a vector of longitude, latitude (in decimal degrees) and altitude (in meters) in WSG84}

\item{a1}{first coordinate of a great circle arc:
a vector of longitude, latitude (in decimal degrees) and elevation (in meters) in WSG84}

\item{a2}{second coordinate of a great circle arc:
a vector of longitude, latitude (in decimal degrees) and elevation (in meters) in WSG84}
}
\value{
a WSG84 vector with longitude and latitude (decimal degrees)
}
\description{
The cross-track intersection between the position of a body (i.e. aircraft) and a great circle
(as defined for example by a runway thresholds).
NOTE: elevation/altitude is ignored and assumed zero.
}
\examples{
\dontrun{
# aircraft (longitude, latitude, altitude)
b <- c(8.086135, 49.973942, 6401)
# 07R (longitude, latitude, altitude)
a1 <- c(8.53417, 50.0275, 328)
# 25L
a2 <- c(8.58653, 50.0401, 362)
cross_track_intersection(b, a1, a2)
}
}
\seealso{
Other utilities: 
\code{\link{altitude_azimuth_distance}()}
}
\concept{utilities}
