% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_lon2n_E.R
\name{lat_lon2n_E}
\alias{lat_lon2n_E}
\title{Convert latitude and longitude to n-vector.}
\usage{
lat_lon2n_E(latitude, longitude)
}
\arguments{
\item{latitude}{Geodetic latitude (rad)}

\item{longitude}{Geodetic longitude (rad)}
}
\value{
n-vector decomposed in E (3x1 vector) (no unit)
}
\description{
Convert latitude and longitude to n-vector.
}
\examples{
\dontrun{
n_E <- lat_lon2n_E(latitude, longitude)
}
}
