% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_E2lat_lon.R
\name{n_E2lat_lon}
\alias{n_E2lat_lon}
\title{Convert n-vector to latitude and longitude.}
\usage{
n_E2lat_lon(n_E)
}
\arguments{
\item{n_E}{n-vector decomposed in E (3x1 vector) (no unit)}
}
\value{
A vector of geodetic latitude and longitude (radians)
}
\description{
Convert n-vector to latitude and longitude.
}
