% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_E2R_EN.R
\name{n_E2R_EN}
\alias{n_E2R_EN}
\title{Finds the rotation matrix R_EN from n-vector.}
\usage{
n_E2R_EN(n_E)
}
\arguments{
\item{n_E}{n-vector decomposed in E (3x1 vector) (unitless)}
}
\value{
The resulting rotation matrix (direction cosine matrix)
}
\description{
Finds the rotation matrix R_EN from n-vector.
}
