% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_EB_E2p_EB_E.R
\name{n_EB_E2p_EB_E}
\alias{n_EB_E2p_EB_E}
\title{Convert n-vector to Cartesian position vector in meters.}
\usage{
n_EB_E2p_EB_E(n_EB_E, z_EB = 0, a = 6378137, f = 1/298.257223563)
}
\arguments{
\item{n_EB_E}{n-vector of position B, decomposed in E (3x1 vector) (no unit)}

\item{z_EB}{Depth of system B, relative to the ellipsoid (z_EB = -height) (m, default 0)}

\item{a}{Semi-major axis of the Earth ellipsoid (m, default [WGS-84] 6378137)}

\item{f}{Flattening of the Earth ellipsoid (no unit, default [WGS-84] 1/298.257223563)}
}
\value{
Cartesian position vector from E to B, decomposed in E (3x1 vector) (m)
}
\description{
The function converts the position of B (typically body) relative to E (typically Earth),
the n-vector \code{n_EB_E} to cartesian position vector ("ECEF-vector"), \code{p_EB_E},
in meters.
}
\examples{
\\donotrun{
n_EB_E2p_EB_E(n_EB_E)
}

}
