% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Ee.R
\name{R_Ee}
\alias{R_Ee}
\title{Select axes of the coordinate frame E.}
\usage{
R_Ee(axes = "e")
}
\arguments{
\item{axes}{Either 'e' or 'E'
* 'e': z-axis points to the North Pole along the Earth's rotation axis,
  x-axis points towards the point where latitude = longitude = 0.
  This choice is very common in many fields.
  'E': x-axis points to the North Pole along the Earth's rotation axis,
  y-axis points towards longitude +90deg (east) and latitude = 0.
  (the yz-plane coincides with the equatorial plane).
  This choice of axis ensures that at zero latitude and longitude,
  frame N (North-East-Down) has the same orientation as frame E.
  If roll/pitch/yaw are zero, also frame B (forward-starboard-down)
  has this orientation. In this manner, the axes of frame E is
  chosen to correspond with the axes of frame N and B.
  The functions in this library originally used this option.}
}
\value{
rotation matrix defining the axes of the coordinate frame E as
        described in Table 2 in Gade (2010)
}
\description{
This file controls the axes of the coordinate frame E (Earth-Centred,
Earth-Fixed, ECEF) used by the other files in this library
}
\details{
There are two choices of E-axes that are described in Table 2 in Gade (2010):

* e: z-axis points to the North Pole and x-axis points to the point where
     latitude = longitude = 0. This choice is very common in many fields.

* E: x-axis points to the North Pole, y-axis points towards longitude +90deg
     (east) and latitude = 0. This choice of axis directions ensures
     that at zero latitude and longitude, N (North-East-Down) has the
     same orientation as E. If roll/pitch/yaw are zero, also B (Body,
     forward, starboard, down) has this orientation. In this manner, the
     axes of E is chosen to correspond with the axes of N and B.
}
\examples{
\dontrun{
R_Ee()
}
}
