% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EN2n_E.R
\name{R_EN2n_E}
\alias{R_EN2n_E}
\title{Find n-vector from R_E}
\usage{
R_EN2n_E(R_EN)
}
\arguments{
\item{R_EN}{Rotation matrix (direction cosine matrix) (no unit)}
}
\value{
n-vector decomposed in E (3x1 vector) (no unit)
}
\description{
Find n-vector from R_E
}
\examples{
\dontrun{
R_EN2n_E(R_EN)
}
}
