% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad}
\alias{rad}
\title{Convert angle in degrees to radians.}
\usage{
rad(degrees)
}
\arguments{
\item{degrees}{angle in degrees.}
}
\value{
angle in radians
}
\description{
Convert angle in degrees to radians.
}
\examples{
rad(30)
}
\seealso{
\code{\link{deg}}.

Other helpers: 
\code{\link{deg}()},
\code{\link{unit}()}
}
\concept{helpers}
