% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unit}
\alias{unit}
\title{Make input vector unit length, i.e. norm == 1}
\usage{
unit(vector)
}
\arguments{
\item{vector}{a vector}
}
\value{
a unit length vector
}
\description{
Make input vector unit length, i.e. norm == 1
}
\examples{
unit(c(1,2,3))

}
\seealso{
Other helpers: 
\code{\link{deg}()},
\code{\link{rad}()}
}
\concept{helpers}
