% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg}
\alias{deg}
\title{Convert angle in radians to degrees}
\usage{
deg(radians)
}
\arguments{
\item{radians}{angle in radians.}
}
\value{
angle in degrees.
}
\description{
Convert angle in radians to degrees
}
\examples{
deg(pi/2)
}
\seealso{
\code{\link{rad}}.

Other helpers: 
\code{\link{rad}()},
\code{\link{unit}()}
}
\concept{helpers}
