% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_lon2n_E.R
\name{lat_lon2n_E}
\alias{lat_lon2n_E}
\title{Convert (geodetic) latitude and longitude to n-vector}
\usage{
lat_lon2n_E(latitude, longitude)
}
\arguments{
\item{latitude}{Geodetic latitude (rad)}

\item{longitude}{Geodetic longitude (rad)}
}
\value{
n-vector decomposed in E (3x1 vector) (no unit)
}
\description{
Convert (geodetic) latitude and longitude to n-vector
}
\examples{
lat_lon2n_E(rad(1), rad(2))

}
\references{
Kenneth Gade \href{https://www.navlab.net/Publications/A_Nonsingular_Horizontal_Position_Representation.pdf}{A Nonsingular Horizontal Position Representation}.
\emph{The Journal of Navigation}, Volume 63, Issue 03, pp 395-417, July 2010.
}
\seealso{
\code{\link{n_E2lat_lon}}.
}
