% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_E2lat_lon.R
\name{n_E2lat_lon}
\alias{n_E2lat_lon}
\title{Convert n-vector to latitude and longitude}
\usage{
n_E2lat_lon(n_E)
}
\arguments{
\item{n_E}{n-vector decomposed in E (3x1 vector) (no unit)}
}
\value{
A vector of geodetic latitude and longitude (rad)
}
\description{
Convert n-vector to latitude and longitude
}
\examples{
n_E2lat_lon(c(1, 0, 0))

}
\references{
Kenneth Gade \href{https://www.navlab.net/Publications/A_Nonsingular_Horizontal_Position_Representation.pdf}{A Nonsingular Horizontal Position Representation}.
\emph{The Journal of Navigation}, Volume 63, Issue 03, pp 395-417, July 2010.
}
\seealso{
\code{\link{lat_lon2n_E}}.
}
