% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvctr-package.R
\docType{package}
\name{nvctr}
\alias{nvctr}
\alias{nvctr-package}
\title{nvctr: A package for computating .}
\description{
The `nvctr` package provides functions to calculate geographical
positions for  both the ellipsoidal and spherical Earth models.
}
