% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvctr-package.R
\docType{package}
\name{nvctr}
\alias{nvctr}
\alias{nvctr-package}
\title{nvctr: non-singular geographical position calculations}
\description{
nvctr provides functions to calculate geographical
positions for  both the ellipsoidal and spherical Earth models.
}
\references{
Kenneth Gade \href{www.navlab.net/Publications/A_Nonsingular_Horizontal_Position_Representation.pdf}{A Nonsingular Horizontal Position Representation}.
\emph{The Journal of Navigation}, Volume 63, Issue 03, pp 395-417, July 2010.
}
